change tg_id column type from Integer to BigInteger in clients model

+++ bot/database/clients/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import BigInteger, Column, Integer, String, ForeignKey
 
 from bot.database.database import Base
 
@@ -6,6 +6,6 @@ class Client(Base):
     __tablename__ = "clients"
 
     id = Column(Integer, primary_key=True)
-    tg_id = Column(Integer, nullable=False, unique=True)
+    tg_id = Column(BigInteger, nullable=False, unique=True)
     table_id = Column(ForeignKey("tables.id", ondelete="CASCADE"), nullable=True)
     username = Column(String(32), nullable=False)
